+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - ERASE, PARSE, TRUE, FALSE
 ### Changed
  - S" max string length is reduced to 255 characters.
+ - Multi-line S" is no longer allowed.
+ - MML string definitions are now done using the MML" word.
  - Changed REFILL to Forth standard behavior: Fill the input buffer from the input source, returning true if successful.
  - #> (the pictured numeric output string buffer) now uses its own buffer, separate from HERE.
  - Do not print "ok" while compiling. Makes it easier to re-enter multi-line word definitions in interpreter.